third_party/blink/web_tests/fast/overflow/hit-test-overflow-hidden-with-box-shadow-expected.txt

This test checks that div block should not get events on clicking the shadow outside div block.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS successfullyParsed is true

TEST COMPLETE


Focus should remain in the textarea
PASS document.activeElement.tagName is "TEXTAREA"

Focus should move to the body
PASS document.activeElement.tagName is "BODY"

Focus should remain in the textarea
PASS document.activeElement.tagName is "TEXTAREA"

Focus should move to the body
PASS document.activeElement.tagName is "BODY"
